Split up agents and OpenAI-interface-compatible agents
Split up our normal agents from OpenAI-interface-compatible agents. The
interface compatible agents let you make calls to
`agent.chat.completions.create`, which the normal agents let you call
`agent.create_completion`.

We put the OpenAI-compatible agents inside the
`src/fixpoint/agents/oai/` folder.

I made this split for two reasons:

1. In case we have drift between what's supported for OpenAI
   interface compatibility and what's supported as a Fixpoint agent
2. Because I wanted to put the `OpenAIAgent` and `OpenAI` classes in
   different files to link to them in docs.
