+++ src/fixpoint/agents/__init__.py
@@ -3,7 +3,8 @@ This is the agents module.
 """
 
 from .protocol import BaseAgent
-from .openai import OpenAI, OpenAIAgent
+from .openai import OpenAIAgent
 from ._shared import CacheMode
+from . import oai
 
-__all__ = ["BaseAgent", "OpenAIAgent", "OpenAI", "CacheMode"]
+__all__ = ["BaseAgent", "OpenAIAgent", "CacheMode", "oai"]
